+++ src/render/surface.rs
@@ -77,6 +77,10 @@ impl Renderer {
         self.frame.resize(self.target.byte_len(), 0);
         self.fxaa_scratch.resize(self.target.byte_len(), 0);
         self.bloom_scratch.resize(self.target.byte_len(), 0);
+        self.oit_scratch.resize(
+            self.target.pixel_len(),
+            super::cpu::OitAccumPixel::default(),
+        );
         if let Some(linear_frame) = &mut self.linear_frame {
             linear_frame.resize(self.target.pixel_len(), Color::BLACK);
         }
@@ -149,6 +153,10 @@ impl Renderer {
         self.frame.resize(self.target.byte_len(), 0);
         self.fxaa_scratch.resize(self.target.byte_len(), 0);
         self.bloom_scratch.resize(self.target.byte_len(), 0);
+        self.oit_scratch.resize(
+            self.target.pixel_len(),
+            super::cpu::OitAccumPixel::default(),
+        );
         if let Some(linear_frame) = &mut self.linear_frame {
             linear_frame.resize(self.target.pixel_len(), Color::BLACK);
         }